+++ cmd/newrelic/command.go
@@ -1,24 +1,179 @@
 package main
 
 import (
+	"errors"
+	"fmt"
 	"os"
+	"strconv"
 
+	"github.com/jedib0t/go-pretty/v6/text"
+	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
+	"github.com/newrelic/newrelic-cli/internal/client"
+	"github.com/newrelic/newrelic-cli/internal/credentials"
 	"github.com/newrelic/newrelic-cli/internal/output"
 	"github.com/newrelic/newrelic-cli/internal/utils"
+	"github.com/newrelic/newrelic-client-go/newrelic"
+	"github.com/newrelic/newrelic-client-go/pkg/accounts"
+	"github.com/newrelic/newrelic-client-go/pkg/nerdgraph"
 )
 
 var outputFormat string
 var outputPlain bool
 
+const defaultProfileName string = "default"
+
 // Command represents the base command when called without any subcommands
 var Command = &cobra.Command{
+	PersistentPreRun:  initializeCLI,
 	Use:               appName,
 	Short:             "The New Relic CLI",
 	Long:              `The New Relic CLI enables users to perform tasks against the New Relic APIs`,
 	Version:           version,
 	DisableAutoGenTag: true, // Do not print generation date on documentation
+}
+
+func initializeCLI(cmd *cobra.Command, args []string) {
+	initializeProfile()
+}
+
+func initializeProfile() {
+	var accountID int
+	var region string
+	var licenseKey string
+	var err error
+
+	credentials.WithCredentials(func(c *credentials.Credentials) {
+		if c.DefaultProfile != "" {
+			err = errors.New("default profile already exists, not attempting to initialize")
+			return
+		}
+
+		apiKey := os.Getenv("NEW_RELIC_API_KEY")
+		envAccountID := os.Getenv("NEW_RELIC_ACCOUNT_ID")
+
+		region = os.Getenv("NEW_RELIC_REGION")
+		licenseKey = os.Getenv("NEW_RELIC_LICENSE_KEY")
+
+		// If we don't have a personal API key we can't initialize a profile.
+		if apiKey == "" {
+			err = errors.New("api key not provided, not attempting to initialize default profile")
+			return
+		}
+
+		// Default the region to US if it's not in the environment
+		if region == "" {
+			region = "US"
+		}
+
+		// Use the accountID from the environment if we have it.
+		if envAccountID != "" {
+			accountID, err = strconv.Atoi(envAccountID)
+			if err != nil {
+				err = fmt.Errorf("couldn't parse account ID: %s", err)
+				return
+			}
+		}
+
+		// We should have an API key by this point, initialize the client.
+		client.WithClient(func(nrClient *newrelic.NewRelic) {
+			// If we still don't have an account ID try to look one up from the API.
+			if accountID == 0 {
+				accountID, err = fetchAccountID(nrClient)
+				if err != nil {
+					return
+				}
+			}
+
+			if licenseKey == "" {
+				// We should have an account ID by now, so fetch the license key for it.
+				licenseKey, err = fetchLicenseKey(nrClient, accountID)
+				if err != nil {
+					return
+				}
+			}
+
+			if !hasProfileWithDefaultName(c.Profiles) {
+				p := credentials.Profile{
+					Region:     region,
+					APIKey:     apiKey,
+					AccountID:  accountID,
+					LicenseKey: licenseKey,
+				}
+
+				err = c.AddProfile(defaultProfileName, p)
+				if err != nil {
+					return
+				}
+
+				log.Infof("profile %s added", text.FgCyan.Sprint(defaultProfileName))
+			}
+
+			if len(c.Profiles) == 1 {
+				err = c.SetDefaultProfile(defaultProfileName)
+				if err != nil {
+					err = fmt.Errorf("error setting %s as the default profile: %s", text.FgCyan.Sprint(defaultProfileName), err)
+					return
+				}
+
+				log.Infof("setting %s as default profile", text.FgCyan.Sprint(defaultProfileName))
+			}
+		})
+	})
+
+	if err != nil {
+		log.Debugf("couldn't initialize default profile: %s", err)
+	}
+}
+
+func hasProfileWithDefaultName(profiles map[string]credentials.Profile) bool {
+	for profileName := range profiles {
+		if profileName == defaultProfileName {
+			return true
+		}
+	}
+
+	return false
+}
+
+func fetchLicenseKey(client *newrelic.NewRelic, accountID int) (string, error) {
+	query := ` query($accountId: Int!) { actor { account(id: $accountId) { licenseKey } } }`
+
+	variables := map[string]interface{}{
+		"accountId": accountID,
+	}
+
+	resp, err := client.NerdGraph.Query(query, variables)
+	if err != nil {
+		return "", err
+	}
+
+	queryResp := resp.(nerdgraph.QueryResponse)
+	actor := queryResp.Actor.(map[string]interface{})
+	account := actor["account"].(map[string]interface{})
+	licenseKey := account["licenseKey"].(string)
+
+	return licenseKey, nil
+}
+
+// fetchAccountID will try and retrieve an account ID for the given user.  If it
+// finds more than one account it will returrn an error.
+func fetchAccountID(client *newrelic.NewRelic) (int, error) {
+	params := accounts.ListAccountsParams{
+		Scope: &accounts.RegionScopeTypes.IN_REGION,
+	}
+
+	accounts, err := client.Accounts.ListAccounts(params)
+	if err != nil {
+		return 0, err
+	}
+
+	if len(accounts) == 1 {
+		return accounts[0].ID, nil
+	}
+
+	return 0, errors.New("more than one account found")
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.