+++ internal/client/client_integration_test.go
@@ -31,7 +31,12 @@ func TestApplyOverrides(t *testing.T) {
 	assert.NoError(t, err)
 
 	// Create an initial profile to work with
-	err = c.AddProfile("testCase1", "us", "apiKeyGoesHere", "insightsInsertKeyGoesHere")
+	testProfile := credentials.Profile{
+		Region:            "us",
+		APIKey:            "apiKeyGoesHere",
+		InsightsInsertKey: "insightsInsertKeyGoesHere",
+	}
+	err = c.AddProfile("testCase1", testProfile)
 	assert.NoError(t, err)
 	p := c.Profiles["testCase1"]
 	assert.NotNil(t, p)