+++ scripts/install.sh
@@ -3,7 +3,7 @@
 # Install the New Relic CLI.
 # https://github.com/newrelic/newrelic-cli
 #
-# Dependencies: curl, cut, tar
+# Dependencies: curl, cut, tar, gzip
 #
 # The version to install and the binary location can be passed in via VERSION and DESTDIR respectively.
 #
@@ -11,6 +11,20 @@
 set -o errexit
 
 echo "Starting installation."
+
+# Determine release filename. This can be expanded with CPU arch in the future.
+if [ "$(uname)" == "Linux" ]; then
+	OS="Linux"
+elif [ "$(uname)" == "Darwin" ]; then
+	OS="Darwin"
+else
+	echo "This operating system is not supported."
+	exit 1
+fi
+
+for x in curl cut tar gzip; do
+    which $x > /dev/null || (echo "Unable to continue.  Please install $x before proceed."; exit 1)
+done
 
 # GitHub's URL for the latest release, will redirect.
 LATEST_URL="https://github.com/newrelic/newrelic-cli/releases/latest"
@@ -32,16 +46,6 @@ function error {
 }
 
 trap error ERR
-
-# Determine release filename. This can be expanded with CPU arch in the future.
-if [ "$(uname)" == "Linux" ]; then
-	OS="Linux"
-elif [ "$(uname)" == "Darwin" ]; then
-	OS="Darwin"
-else
-	echo "This operating system is not supported."
-	exit 1
-fi
 
 RELEASE_URL="https://github.com/newrelic/newrelic-cli/releases/download/v${VERSION}/newrelic-cli_${VERSION}_${OS}_x86_64.tar.gz"